Most of new internals for indexing are in place
