+++ src/axis_interface.jl
@@ -1,156 +1,5 @@
 
 """
-    Axis(k[, v=OneTo(length(k))])
-
-Subtypes of `AbstractAxis` that maps keys to values. The first argument specifies
-the keys and the second specifies the values. If only one argument is specified
-then the values span from 1 to the length of `k`.
-
-## Examples
-
-The value for all of these is the same.
-```jldoctest axis_examples
-julia> using AxisIndices
-
-julia> x = Axis(2.0:11.0, 1:10)
-Axis(2.0:1.0:11.0 => 1:10)
-
-julia> y = Axis(2.0:11.0)  # when only one argument is specified assume it's the keys
-Axis(2.0:1.0:11.0 => Base.OneTo(10))
-
-julia> z = Axis(1:10)
-Axis(1:10 => Base.OneTo(10))
-```
-
-Standard indexing returns the same values
-```jldoctest axis_examples
-julia> x[2]
-2
-
-julia> x[2] == y[2] == z[2]
-true
-
-julia> x[1:2]
-Axis(2.0:1.0:3.0 => 1:2)
-
-julia> y[1:2]
-Axis(2.0:1.0:3.0 => 1:2)
-
-julia> z[1:2]
-Axis(1:2 => 1:2)
-
-julia> x[1:2] == y[1:2] == z[1:2]
-true
-```
-
-Functions that return `true` or `false` may be used to search the keys for their
-corresponding index. The following is equivalent to the previous example.
-```jldoctest axis_examples
-julia> x[==(3.0)]
-2
-
-julia> x[==(3.0)] ==       # 3.0 is the 2nd key of x
-       y[isequal(3.0)] ==  # 3.0 is the 2nd key of y
-       z[==(2)]            # 2 is the 2nd key of z
-true
-
-julia> x[<(4.0)]  # all keys less than 4.0 are 2.0:3.0 which correspond to values 1:2
-Axis(2.0:1.0:3.0 => 1:2)
-
-julia> y[<=(3.0)]  # all keys less than or equal to 3.0 are 2.0:3.0 which correspond to values 1:2
-Axis(2.0:1.0:3.0 => 1:2)
-
-julia> z[<(3)]  # all keys less than or equal to 3 are 1:2 which correspond to values 1:2
-Axis(1:2 => 1:2)
-
-julia> x[<(4.0)] == y[<=(3.0)] == z[<(3)]
-true
-```
-Notice that `==` returns a single value instead of a collection of all elements
-where the key was found to be true. This is because all keys must be unique so
-there can only ever be one element returned.
-"""
-struct Axis{K,I,Ks,Inds<:AbstractUnitRange{I}} <: AbstractAxis{K,I}
-    keys::Ks
-    parent_indices::Inds
-
-    function Axis{K,I,Ks,Inds}(
-        ks::Ks,
-        inds::Inds,
-        check_unique::Bool=true,
-        check_length::Bool=true
-    ) where {K,I,Ks<:AbstractVector{K},Inds<:AbstractUnitRange{I}}
-        check_unique && check_axis_unique(ks, inds)
-        check_length && check_axis_length(ks, inds)
-        return new{K,I,Ks,Inds}(ks, inds)
-    end
-
-    function Axis{K,V,Ks,Vs}(x::AbstractUnitRange{<:Integer}) where {K,V,Ks,Vs}
-        if x isa Ks
-            if x isa Vs
-                return Axis{K,V,Ks,Vs}(x, x)
-            else
-                return  Axis{K,V,Ks,Vs}(x, Vs(x))
-            end
-        else
-            if x isa Vs
-                return Axis{K,V,Ks,Vs}(Ks(x), x)
-            else
-                return  Axis{K,V,Ks,Vs}(Ks(x), Vs(x))
-            end
-        end
-    end
-
-    # Axis{K,I}
-    function Axis{K,I}() where {K,I}
-        return new{K,I,Vector{K},OneToMRange{I}}(Vector{K}(),OneToMRange{I}(0))
-    end
-
-    function Axis{K,I,Ks,Inds}(axis::AbstractAxis) where {K,I,Ks,Inds}
-        return Axis{K,I,Ks,Inds}(Ks(keys(axis)), Inds(parentindices(axis)), false, false)
-    end
-
-    function Axis{K,I,Ks,Inds}(axis::Axis{K,I,Ks,Inds}) where {K,I,Ks,Inds}
-        if can_change_size(axis)
-            return copy(axis)
-        else
-            return axis
-        end
-    end
-
-    function Axis(axis::AbstractAxis)
-        if can_change_size(axis)
-            return axis
-        else
-            return copy(axis)
-        end
-    end
-
-    Axis{K}() where {K} = Axis{K,Int}()
-
-    Axis() = Axis{Any}()
-
-    Axis(x::Pair) = Axis(x.first, x.second)
-
-    function Axis(ks, inds, check_unique::Bool=true, check_length::Bool=true)
-        return Axis{eltype(ks),eltype(inds),typeof(ks),typeof(inds)}(ks, inds, check_unique, check_length)
-    end
-
-    function Axis(ks, check_unique::Bool=true)
-        if can_change_size(ks)
-            return Axis(ks, OneToMRange(length(ks)), check_unique, false)
-        else
-            len = known_length(ks)
-            if len isa Nothing
-                return return Axis(ks, OneTo(length(ks)), check_unique, false)
-            else
-                return Axis(ks, OneToSRange(len), false)
-            end
-        end
-    end
-end
-
-"""
     SimpleAxis(v)
 
 Povides an `AbstractAxis` interface for any `AbstractUnitRange`, `v `. `v` will
@@ -236,33 +85,11 @@ struct SimpleAxis{I,Inds<:AbstractUnitRange{I}} <: AbstractAxis{I,I}
     SimpleAxis(stop::Integer) = SimpleAxis(OneTo(stop))
 end
 
-Base.IndexStyle(::Type{T}) where {T<:Axis} = IndexAxis()
-function unsafe_reconstruct(::IndexAxis, axis, arg, inds)
-    if is_key(axis, arg) && (arg isa AbstractVector)
-        ks = arg
-    else
-        ks = @inbounds(getindex(keys(axis), inds))
-    end
-    return Axis(ks, unsafe_reconstruct(parentindices(axis), arg, inds), false, false)
-end
-
-function unsafe_reconstruct(::IndexAxis, axis, inds)
-    return Axis(
-        @inbounds(getindex(keys(axis), inds)),
-        unsafe_reconstruct(parentindices(axis), arg, inds),
-        false,
-        false
-    )
-end
-
 # Axis interface
-Base.parentindices(axis::Axis) = getfield(axis, :parent_indices)
 Base.parentindices(axis::SimpleAxis) = getfield(axis, :parent_indices)
 
-ArrayInterface.parent_type(::Type{T}) where {Inds,T<:Axis{<:Any,<:Any,<:Any,Inds}} = Inds
 ArrayInterface.parent_type(::Type{T}) where {Inds,T<:SimpleAxis{<:Any,Inds}} = Inds
 
-Base.keys(axis::Axis) = getfield(axis, :keys)
 Base.keys(axis::SimpleAxis) = parentindices(axis)
 
 # :resize_first!, :resize_last! don't need to define these ones b/c non mutating ones are only
@@ -369,24 +196,6 @@ function _popfirst!(axis::Axis)
 end
 
 _popfirst!(axis::AbstractAxis) = popfirst!(parentindices(axis))
-
-#= TODO implement this when part of ArrayInterface
-function StaticRanges.popfirst(axis::AbstractAxis)
-    if is_indices_axis(axis)
-        return unsafe_reconstruct(axis, popfirst(indices(axis)))
-    else
-        return unsafe_reconstruct(axis, popfirst(keys(axis)), popfirst(indices(axis)))
-    end
-end
-
-function StaticRanges.pop(axis::AbstractAxis)
-    if is_indices_axis(axis)
-        return unsafe_reconstruct(axis, pop(indices(axis)))
-    else
-        return unsafe_reconstruct(axis, pop(keys(axis)), pop(indices(axis)))
-    end
-end
-=#
 
 # TODO check for existing key first
 function push_key!(axis::AbstractAxis, key)
@@ -548,15 +357,6 @@ function Base.summary(io::IO, a::AbstractAxis)
     return print(io, "$(length(a))-element $(typeof(a).name)($(keys(a)) => $(values(a)))")
 end
 
-#=
-function reverse_keys(axis::AbstractAxis, newinds::AbstractUnitRange)
-    if is_indices_axis(axis)
-        return unsafe_reconstruct(reverse(keys(axis)), newinds, false)
-    else
-        return similar(axis, reverse(keys(axis)), newinds, false)
-    end
-end
-=#
 function reverse_keys(axis::AbstractAxis, newinds::AbstractUnitRange)
     return Axis(reverse(keys(axis)), newinds, false, false)
 end
