+++ src/deprecations.jl
@@ -1,0 +1,2 @@
+
+@deprecate axes_keys(x) keys.(axes(x))
