+++ test/Axes/Axes.jl
@@ -1,6 +1,6 @@
 
 include("Axis.jl")
-include("MetaAxis.jl")
-include("indexing.jl")
-include("to_axis.jl")
+#include("MetaAxis.jl")
+#include("indexing.jl")
+#include("to_axis.jl")
